+++ imports/state/reducers/quote/chargesReducers.js
@@ -54,7 +54,15 @@ export const chargeLines = (
   }
   return newState.map((chargeLine) => {
     const res = { ...chargeLine };
-    res.amount = (chargeLine.units || 0) * (chargeLine.unitPrice || 0);
+    let minimumAmount = 0;
+    if (chargeLine.selectedRate !== 'custom') {
+      minimumAmount = chargeLine.applicableSellRates[chargeLine.selectedRate].minimumAmount || 0;
+    }
+    let amount = (chargeLine.units || 0) * (chargeLine.unitPrice || 0);
+    if (amount < minimumAmount) {
+      amount = minimumAmount;
+    }
+    res.amount = amount;
     if (chargeLine.currency === parentState.currency) {
       res.finalAmount = res.amount;
     } else if (