+++ imports/ui/editors/EditQuoteChargesConnect.js
@@ -9,22 +9,8 @@ import {
   setQuoteCurrency,
 } from '../../state/actions/quoteActions';
 import { loadEmail } from '../../state/actions/emailActions';
-import { getAmount } from '../quoteUtils';
 
-const mapStateToProps = (state) => {
-  const quote = state.quote;
-  return {
-    quote: {
-      ...quote,
-      charges: {
-        ...quote.charges,
-        chargeLines: quote.charges.chargeLines.map(
-          chargeLine => ({ ...chargeLine, amount: getAmount(chargeLine) }),
-        ),
-      },
-    },
-  };
-};
+const mapStateToProps = state => ({ quote: state.quote });
 
 const mapDispatchToProps = dispatch => ({
   onLoad: quote => dispatch(loadQuote(quote)),